Cleanup the test support templates
